+++ programs/sdl2test4/sago/SagoSprite.hpp
@@ -39,7 +39,7 @@ public:
      * @param x Place to draw the sprite
      * @param y Place to draw the sprite
      */
-	void Draw(SDL_Renderer* target, Sint32 frameTime, float x, float y) const;
+	void Draw(SDL_Renderer* target, Sint32 frameTime, int x, int y) const;
 	/**
 	 * Draws part of the sprite to a given render window
      * @param target The render window to draw on
@@ -48,7 +48,7 @@ public:
      * @param y Place to draw the sprite
      * @param part the part of the sprite that should be drawn. 
      */
-	void Draw(SDL_Renderer* target, Sint32 frameTime, float x, float y, const SDL_Rect& part) const;
+	void Draw(SDL_Renderer* target, Sint32 frameTime, int x, int y, const SDL_Rect& part) const;
 	/**
 	 * Draws the wprite to the given renderer but makes sure to not draw outside th bounds given
 	 * @param target The render window to draw on
@@ -57,7 +57,7 @@ public:
 	 * @param y Place to draw the sprite
 	 * @param bounds A recagular area that we must not draw outside.
 	 */
-	void DrawBounded(SDL_Renderer* target, Sint32 frameTime, float x, float y, const SDL_Rect& bounds) const;
+	void DrawBounded(SDL_Renderer* target, Sint32 frameTime, int x, int y, const SDL_Rect& bounds) const;
 	/**
 	 * Set a different origin. Normally it is the top left cornor. But in some cases you might want to center the origin or tranform it for other reasons
      * @param newOrigin the coordinates that should be the new origin. Call with {0,0} to reset to default
